requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
alembic==1.14.0
pymysql==1.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1
python-multipart==0.0.17
email-validator==2.2.0
google-generativeai==0.8.3
openai>=1.0.0
pdf2image>=1.17.0
Pillow>=10.0.0
